+++ CsharpExtrasTest/Dictionary/Collection/SetValuedDictionaryTest.cs
@@ -4,19 +4,36 @@ using NUnit.Framework;
 using System;
 using System.Collections.Generic;
 using System.Text;
+using CsharpExtras.Dictionary.Collection;
 
 namespace CsharpExtrasTest.Dictionary
 {
     [TestFixture]
-    class MultiValueMapTest
+    class SetValuedDictionaryTest
     {
+        [Test]
+        [Category("Unit")]
+        public void GivenEmptyMapWhenTwoItemsAddedToSameKeyThenBothItemsAreInSetForThatKey()
+        {
+            ISetValuedDictionary<string, int> setDictionary = new SetValuedDictionaryImpl<string, int>();
+            string key = "bob";
+            Assert.IsFalse(setDictionary.ContainsKey(key), string.Format(
+                "Given: expected test to start with empty map entry for key {0}, but map appears to contain value for that key."
+                , key));
+            const int value1 = 1;
+            setDictionary.Add(key, value1);
+            const int value2 = 2;
+            setDictionary.Add(key, value2);
+            Assert.AreEqual(setDictionary[key], new int[] { value1, value2 });
+        }
+
         [Test]
         [Category("Unit")]
         public void Given_TwoMaps_When_DifferentNumberOfElementsAdded_Then_MapsAreNotEqual()
         {
             //Assemble
-            IMultiValueMap<int, string> map1 = new MultiValueMapImpl<int, string>();
-            IMultiValueMap<int, string> map2 = new MultiValueMapImpl<int, string>();
+            ISetValuedDictionary<int, string> map1 = new SetValuedDictionaryImpl<int, string>();
+            ISetValuedDictionary<int, string> map2 = new SetValuedDictionaryImpl<int, string>();
 
             //Act
             map1.Add(1, "Unity");
@@ -37,8 +54,8 @@ namespace CsharpExtrasTest.Dictionary
         public void Given_TwoMaps_When_DifferentElementsAdded_Then_MapsAreNotEqual()
         {
             //Assemble
-            IMultiValueMap<int, string> map1 = new MultiValueMapImpl<int, string>();
-            IMultiValueMap<int, string> map2 = new MultiValueMapImpl<int, string>();
+            ISetValuedDictionary<int, string> map1 = new SetValuedDictionaryImpl<int, string>();
+            ISetValuedDictionary<int, string> map2 = new SetValuedDictionaryImpl<int, string>();
 
             //Act
             map1.Add(1, "Unity");
@@ -60,8 +77,8 @@ namespace CsharpExtrasTest.Dictionary
         public void Given_TwoMaps_When_SameElementsAddedInDifferentOrder_Then_MapsAreEqual()
         {
             //Assemble
-            IMultiValueMap<int, string> map1 = new MultiValueMapImpl<int, string>();
-            IMultiValueMap<int, string> map2 = new MultiValueMapImpl<int, string>();
+            ISetValuedDictionary<int, string> map1 = new SetValuedDictionaryImpl<int, string>();
+            ISetValuedDictionary<int, string> map2 = new SetValuedDictionaryImpl<int, string>();
 
             //Act
             map1.Add(1, "Unity");
@@ -75,7 +92,7 @@ namespace CsharpExtrasTest.Dictionary
             map2.Add(1, "One");
 
             //Assert
-            Assert.IsTrue(map1.Equals(map2));
+            Assert.IsTrue(map1.DictEquals(map2));
         }
 
         [Test]
@@ -83,7 +100,7 @@ namespace CsharpExtrasTest.Dictionary
         public void GivenMap_WhenTransform_ThenResultIsOfCorrectLengthAndContainsExpectedMappings()
         {
             //Arrange
-            IMultiValueMap<int, string> testMap = new MultiValueMapImpl<int, string>();
+            ISetValuedDictionary<int, string> testMap = new SetValuedDictionaryImpl<int, string>();
             testMap.Add(1, "Unity");
             testMap.Add(1, "One");
             testMap.Add(2, "Two");
@@ -92,7 +109,7 @@ namespace CsharpExtrasTest.Dictionary
             testMap.Add(3, "Three");
             
             //Act
-            IMultiValueMap<int, int> transformedMap = testMap.TransformValues(s => s.Length);
+            ISetValuedDictionary<int, int> transformedMap = testMap.TransformValues(s => s.Length);
             
             //Assert
             Assert.AreEqual(3, transformedMap.Count, "Transformed map should be the same length as original");
@@ -103,36 +120,10 @@ namespace CsharpExtrasTest.Dictionary
 
         [Test]
         [Category("Unit")]
-        public void Given_TwoMaps_When_TheSameDataIsAddedInDifferentOrders_Then_MapHashCodesMatch()
-        {
-            //Assemble
-            IMultiValueMap<int, string> map1 = new MultiValueMapImpl<int, string>();
-            IMultiValueMap<int, string> map2 = new MultiValueMapImpl<int, string>();
-
-            //Act
-            map1.Add(1, "a");
-            map1.Add(1, "b");
-            map1.Add(1, "c");
-            map1.Add(2, "d");
-            map1.Add(2, "e");
-
-            map2.Add(2, "e");
-            map2.Add(1, "c");
-            map2.Add(1, "a");
-            map2.Add(2, "d");
-            map2.Add(1, "b");
-
-            //Assert
-            Assert.IsTrue(map1.Equals(map2), "GIVEN: Maps should be equal");
-            Assert.AreEqual(map1.GetHashCode(), map2.GetHashCode(), "Map hash codes should match when the maps have the same data");
-        }
-
-        [Test]
-        [Category("Unit")]
         public void Given_OneLargeMap_When_HashCodeGenerated_Then_NoExceptionThrown()
         {
             //Assemble
-            IMultiValueMap<int, string> map = new MultiValueMapImpl<int, string>();
+            ISetValuedDictionary<int, string> map = new SetValuedDictionaryImpl<int, string>();
 
             //Act
             for (int index1 = 0; index1 < 20; index1++)