+++ CHANGELOG.md
@@ -11,9 +11,11 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - [BREAKING] Move the ArrayOrientation enum to a new namespace (#9)
 - [BREAKING] Rename main namespace from `CustomExtensions` to `Extensions` (#9)
 - [BREAKING] The `TryGetSingleton` extension is now only available for collections with elements of type `class` (#49)
+- [BREAKING] The `IMultiValueMap` has been renamed to `ISetValuedDictionary` and its package is different (#53)
 - 2D one-based arrays are now enumerable (#45)
 - Multi-value maps are now comparable using set-equality as the default equals method (#48)
 - Project now treats all warnings as errors, minor changes made to fix all errors (#49)
+- There is now a type `IListValuedDictionary` which is analogous to `ISetValuedDictionary` but for lists vs. sets (#53)
 
 ### Added
 - Support for writing a 1D one-based array to a 2D one-based array (#9)